Extract ActiveUsers to tce-core
